// src/types.ts
export interface FraudData {
  assured_age: number;
  nominee_relation: string;
  occupation: string;
  policy_sum_assured: number;
  premium: number;
  premium_payment_mode: string;
  annual_income: number;
  holder_marital_status: string;
  indiv_requirement_flag: string;
  policy_term: number;
  policy_payment_term: number;
  product_type: string;
  channel: string;
  bank_code: string;
  policy_risk_commencement_date: string;
  date_of_death: string;
  intimation_date: string;
  status: string;
  sub_status: string;
}